Ensure POSIX compliance and fix minor issues

+++ h3client/posix.h
@@ -1,13 +1,12 @@
 #ifndef POSIX_H
 #define POSIX_H
 
-#include <stddef.h>
 #include <string.h>
 
 static inline size_t posix_strnlen(char const *s, size_t n)
 {
   const char *p = memchr(s, 0, n);
-  return p ? p - s : n;
+  return p ? (size_t)(p - s) : n;
 }
 
 #endif
